#87 Base filter

+++ backend/backend/backend/src/Application/Languages/Commands/DeleteLanguage/DeleteLanguageCommand.cs
@@ -11,7 +11,6 @@ namespace backend.Application.Languages.Commands.DeleteLanguage;
 public class DeleteLanguageCommand:IRequest
 {
     // make command immutable 
-  
     public int Id { get; init; }
 }
 public class DeleteLanguageCommandHandler : AsyncRequestHandler<DeleteLanguageCommand>
@@ -26,6 +25,10 @@ public class DeleteLanguageCommandHandler : AsyncRequestHandler<DeleteLanguageCo
     protected override async Task Handle(DeleteLanguageCommand request, CancellationToken cancellationToken)
     {
         var entityTobeDeleted = await _dbContext.Languages.Where(x => x.Id == request.Id).SingleOrDefaultAsync();
+        if(entityTobeDeleted == null)
+        {
+            return;
+        }
         _dbContext.Languages.Remove(entityTobeDeleted);
         await _dbContext.SaveChangesAsync(cancellationToken);
     }
